feat: service for user context created

+++ src/main/java/com/bcarsoftware/avantemove_api/services/IUserService.java
@@ -9,8 +9,8 @@ public interface IUserService {
     public abstract User save(UserDTO userDTO);
     public abstract User login(LoginDTO loginDTO);
     public abstract User logout();
-    public abstract User getById(Integer id);
-    public abstract User update(Integer id, UserDTO userDTO);
+    public abstract User getById(Long id);
+    public abstract User update(Long id, UserDTO userDTO);
     public abstract User updatePassword(RecoveryDTO recoveryDTO);
-    public abstract User delete(Integer id);
+    public abstract User delete(Long id);
 }
